Added a script to grab GOES images from the NOAA web site
This script has everything hard coded, maybe worth modifying it
to take command line parameters in the future.

The script saves images to a temp file initially then renames them
to their original name, to avoid part images if the download aborts.

The script parses dates starting at a start date, and checks every
GOES file that should exist between the start date and now. If that
file does not already exist in the $DESTINATION location then the
script will attempt to download it. In this way the script can be
run repeatedly and will always keep the destination up to date as
best as it can, without redownloading and overwriting files.